+++ core/CHANGELOG.md
@@ -6,6 +6,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 ## 1.5.1 (in progress)
 
 + [#2031](https://github.com/luyadev/luya/issues/2031) Updated LazyLoad for `options` array so that `alt` and `title` attributes can be passed to it.
++ [#2020](https://github.com/luyadev/luya/issues/2020) Fixed generated files of theme/create command.
 
 ## 1.5.0 (19. June 2020)